Created sub component
